mma
git-svn-id: http://10.1.8.14:18080/svn/Local_Business/DAT-Event/Club-Event@3534 5e73b0b3-88fa-41e2-a44a-e6aac4854029

+++ src/main/java/com/dat/event/dto/RequestEventPlanDto.java
@@ -10,7 +10,7 @@ public class RequestEventPlanDto {
 
     private String eventName;
     private String description;
-    private int inChargePerson;
+    private Long inChargePerson;
     private SupportedMember[] supportedMembers;
     private EventTime[] eventTimes;
